backend/requirements.txt
Django==4.1.4
djangorestframework==3.14.0
pandas==1.5.0
python-dotenv==0.21.0
